test: increase the timeout threshold in alpha-loop to 30 seconds
QEMU may be extremely slow when emulating some CPU architectures.
For example, the current 5 seconds timeout elapses in the alpha-loop
test when emulating microblaze on Intel Core i7 860 @2.8GHz. It
is necessary to increase this timeout to at least 15 seconds to
make the test pass in this setup. And in order to have additional
safety headroom, it's better to increase this timeout even more.

QEMU is significantly faster when emulating ARM, MIPS, PowerPC and
other mainstream architectures. So this workaround is not needed for
them.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>